backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Caching==2.0.2
python-dotenv==1.0.0
celery==5.3.1
redis==5.0.0
pandas==2.0.3
fpdf2==2.7.0
requests==2.31.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
